+++ sink/sql/db_proto/sql/postgres/dialect.go
@@ -14,7 +14,6 @@ import (
 	"go.uber.org/zap"
 	"google.golang.org/protobuf/encoding/protojson"
 	"google.golang.org/protobuf/reflect/protoreflect"
-	"google.golang.org/protobuf/types/dynamicpb"
 )
 
 const postgresStaticSql = `
@@ -212,14 +211,14 @@ func (d *DialectPostgres) FullTableName(table *schema.Table) string {
 	return tableName(d.schemaName, table.Name)
 }
 
-func (d *DialectPostgres) AppendInlineFieldValues(fieldValues []any, fd protoreflect.FieldDescriptor, fv protoreflect.Value, dm *dynamicpb.Message) ([]any, error) {
+func (d *DialectPostgres) AppendInlineFieldValues(fieldValues []any, fd protoreflect.FieldDescriptor, fv protoreflect.Value, dm protoreflect.Message) ([]any, error) {
 	if fd.IsList() {
 		// For repeated inline messages, append the list of JSON strings
 		list := fv.List()
 		var jsonStrings []string
 		for j := 0; j < list.Len(); j++ {
-			fm := list.Get(j).Message().Interface().(*dynamicpb.Message)
-			jsonBytes, err := protojson.Marshal(fm)
+			fm := list.Get(j).Message()
+			jsonBytes, err := protojson.Marshal(fm.Interface())
 			if err != nil {
 				return nil, fmt.Errorf("failed to marshal protobuf message to JSON: %w", err)
 			}
@@ -228,8 +227,8 @@ func (d *DialectPostgres) AppendInlineFieldValues(fieldValues []any, fd protoref
 		fieldValues = append(fieldValues, pq.Array(jsonStrings))
 	} else {
 		// For single inline message, append the JSON string
-		fm := fv.Message().Interface().(*dynamicpb.Message)
-		jsonBytes, err := protojson.Marshal(fm)
+		fm := fv.Message()
+		jsonBytes, err := protojson.Marshal(fm.Interface())
 		if err != nil {
 			return nil, fmt.Errorf("failed to marshal protobuf message to JSON: %w", err)
 		}